+++ common/src/main/java/net/bettercombat/logic/WeaponAttributesFallback.java
@@ -1,11 +1,13 @@
 package net.bettercombat.logic;
 
 import net.bettercombat.BetterCombat;
+import net.bettercombat.config.FallbackConfig;
 import net.bettercombat.utils.PatternMatching;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.attribute.EntityAttribute;
 import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.item.Item;
+import net.minecraft.item.RangedWeaponItem;
 import net.minecraft.registry.Registries;
 import net.minecraft.util.Identifier;
 
@@ -14,12 +16,20 @@ public class WeaponAttributesFallback {
         var config = BetterCombat.fallbackConfig.value;
         for(var itemId: Registries.ITEM.getIds()) {
             var item = Registries.ITEM.get(itemId);
-            if (!hasAttributeModifier(item, EntityAttributes.GENERIC_ATTACK_DAMAGE)
-                    || PatternMatching.matches(itemId.toString(), config.blacklist_item_id_regex)) {
+            if (PatternMatching.matches(itemId.toString(), config.blacklist_item_id_regex)) {
                 // Skipping items without attack damage attribute
                 continue;
             }
-            for (var fallbackOption: config.fallback_compatibility) {
+            FallbackConfig.CompatibilitySpecifier[] specifiers = null;
+            if (hasAttributeModifier(item, EntityAttributes.GENERIC_ATTACK_DAMAGE)) {
+                specifiers = config.fallback_compatibility;
+            } else if (item instanceof RangedWeaponItem) {
+                specifiers = config.ranged_weapons;
+            }
+            if (specifiers == null) {
+                continue;
+            }
+            for (var fallbackOption: specifiers) {
                 // If - no registration & matches regex
                 if (WeaponRegistry.getAttributes(itemId) == null
                         && PatternMatching.matches(itemId.toString(), fallbackOption.item_id_regex)) {